feat(209-02): narrow all day-index uint48 to uint32 in DegenerusQuests.sol
- DailyQuest struct field: uint48 day -> uint32 day
- 6 events: day/currentDay params narrowed to uint32
- ~15 local variables: uint48 currentDay -> uint32 currentDay
- 9 private function params: uint48 currentDay -> uint32 currentDay
- _currentQuestDay return type and local: uint48 -> uint32
- Removed intermediate uint48 casts in streak arithmetic
- Updated struct layout comment (48b -> 32b)

+++ contracts/DegenerusQuests.sol
@@ -64,7 +64,7 @@ contract DegenerusQuests is IDegenerusQuests {
 
     /// @notice Emitted when a quest slot is rolled for a new day.
     event QuestSlotRolled(
-        uint48 indexed day,
+        uint32 indexed day,
         uint8 indexed slot,
         uint8 questType,
         uint8 flags,
@@ -74,7 +74,7 @@ contract DegenerusQuests is IDegenerusQuests {
     /// @notice Emitted when player quest progress is updated.
     event QuestProgressUpdated(
         address indexed player,
-        uint48 indexed day,
+        uint32 indexed day,
         uint8 indexed slot,
         uint8 questType,
         uint128 progress,
@@ -84,7 +84,7 @@ contract DegenerusQuests is IDegenerusQuests {
     /// @notice Emitted when a quest slot is completed.
     event QuestCompleted(
         address indexed player,
-        uint48 indexed day,
+        uint32 indexed day,
         uint8 indexed slot,
         uint8 questType,
         uint32 streak,
@@ -96,7 +96,7 @@ contract DegenerusQuests is IDegenerusQuests {
         address indexed player,
         uint16 used,
         uint16 remaining,
-        uint48 currentDay
+        uint32 currentDay
     );
 
     /// @notice Emitted when quest streak is manually increased.
@@ -104,14 +104,14 @@ contract DegenerusQuests is IDegenerusQuests {
         address indexed player,
         uint16 amount,
         uint24 newStreak,
-        uint48 currentDay
+        uint32 currentDay
     );
 
     /// @notice Emitted when quest streak resets due to missed days.
     event QuestStreakReset(
         address indexed player,
         uint24 previousStreak,
-        uint48 currentDay
+        uint32 currentDay
     );
 
     /// @notice Emitted when a player completes the level quest.
@@ -223,7 +223,7 @@ contract DegenerusQuests is IDegenerusQuests {
      *
      * Layout (memory):
      * +-------------+----------+-------+---------+
-     * | day (48b)   | type(8b) | flags | version |
+     * | day (32b)   | type(8b) | flags | version |
      * +-------------+----------+-------+---------+
      *
      * Version Semantics:
@@ -231,7 +231,7 @@ contract DegenerusQuests is IDegenerusQuests {
      * - Player progress is invalidated when version mismatches
      */
     struct DailyQuest {
-        uint48 day;       // Quest day identifier (derived by caller, not block timestamp)
+        uint32 day;       // Quest day identifier (derived by caller, not block timestamp)
         uint8 questType;  // One of the QUEST_TYPE_* constants
         uint8 flags;      // Difficulty flags (HIGH/VERY_HIGH)
         uint24 version;     // Bumped when quest mutates mid-day to reset stale player progress
@@ -422,7 +422,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || quantity == 0 || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -536,7 +536,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || flipCredit == 0 || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -592,7 +592,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || burnAmount == 0 || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -647,7 +647,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || amount == 0 || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -702,7 +702,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || amountWei == 0 || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -769,7 +769,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -916,7 +916,7 @@ contract DegenerusQuests is IDegenerusQuests {
         returns (uint256 reward, uint8 questType, uint32 streak, bool completed)
     {
         DailyQuest[QUEST_SLOT_COUNT] memory quests = activeQuests;
-        uint48 currentDay = _currentQuestDay(quests);
+        uint32 currentDay = _currentQuestDay(quests);
         PlayerQuestState storage state = questPlayerState[player];
         if (player == address(0) || amount == 0 || currentDay == 0) {
             return (0, quests[0].questType, state.streak, false);
@@ -962,7 +962,7 @@ contract DegenerusQuests is IDegenerusQuests {
      */
     function getActiveQuests() external view returns (QuestInfo[2] memory quests) {
         DailyQuest[QUEST_SLOT_COUNT] memory local = _materializeActiveQuestsForView();
-        uint48 currentDay = _currentQuestDay(local);
+        uint32 currentDay = _currentQuestDay(local);
         PlayerQuestState memory emptyState;
         for (uint8 slot; slot < QUEST_SLOT_COUNT; ) {
             (quests[slot], , ) = _questViewData(local[slot], emptyState, slot, currentDay);
@@ -998,7 +998,7 @@ contract DegenerusQuests is IDegenerusQuests {
     {
         DailyQuest[QUEST_SLOT_COUNT] memory local = activeQuests;
         PlayerQuestState memory state = questPlayerState[player];
-        uint48 currentDay = _currentQuestDay(local);
+        uint32 currentDay = _currentQuestDay(local);
         streak = state.streak;
         lastCompletedDay = state.lastCompletedDay;
         for (uint8 slot; slot < QUEST_SLOT_COUNT; ) {
@@ -1022,14 +1022,14 @@ contract DegenerusQuests is IDegenerusQuests {
      */
     function getPlayerQuestView(address player) external view returns (PlayerQuestView memory viewData) {
         DailyQuest[QUEST_SLOT_COUNT] memory local = _materializeActiveQuestsForView();
-        uint48 currentDay = _currentQuestDay(local);
+        uint32 currentDay = _currentQuestDay(local);
         PlayerQuestState memory state = questPlayerState[player];
 
         // Preview streak decay: if player missed days beyond available shields, show 0 streak
         uint32 effectiveStreak = state.streak;
         uint24 anchorDay = state.lastActiveDay != 0 ? state.lastActiveDay : state.lastCompletedDay;
-        if (anchorDay != 0 && currentDay > uint48(anchorDay + 1)) {
-            uint32 missedDays = uint32(currentDay - uint48(anchorDay) - 1);
+        if (anchorDay != 0 && currentDay > anchorDay + 1) {
+            uint32 missedDays = currentDay - anchorDay - 1;
             uint16 shields = questStreakShieldCount[player];
             if (missedDays > uint32(shields)) {
                 effectiveStreak = 0;
@@ -1077,7 +1077,7 @@ contract DegenerusQuests is IDegenerusQuests {
         DailyQuest memory quest,
         PlayerQuestState memory state,
         uint8 slot,
-        uint48 currentDay
+        uint32 currentDay
     ) private view returns (QuestInfo memory info, uint128 progress, bool completed) {
         info = QuestInfo({
             day: quest.day,
@@ -1131,7 +1131,7 @@ contract DegenerusQuests is IDegenerusQuests {
      */
     function _currentDayQuestOfType(
         DailyQuest[QUEST_SLOT_COUNT] memory quests,
-        uint48 currentDay,
+        uint32 currentDay,
         uint8 questType
     ) private pure returns (DailyQuest memory quest, uint8 slotIndex) {
         slotIndex = type(uint8).max; // Sentinel for "not found"
@@ -1233,7 +1233,7 @@ contract DegenerusQuests is IDegenerusQuests {
         uint8 slot,
         uint256 delta,
         uint256 target,
-        uint48 currentDay,
+        uint32 currentDay,
         uint256 mintPrice,
         uint8 handlerQuestType,
         uint256 levelDelta,
@@ -1277,11 +1277,11 @@ contract DegenerusQuests is IDegenerusQuests {
      * @param player Player address for event emission and streak shield lookup.
      * @param currentDay The current quest day.
      */
-    function _questSyncState(PlayerQuestState storage state, address player, uint48 currentDay) private {
+    function _questSyncState(PlayerQuestState storage state, address player, uint32 currentDay) private {
         uint24 prevStreak = state.streak;
         uint24 anchorDay = state.lastActiveDay != 0 ? state.lastActiveDay : state.lastCompletedDay;
-        if (anchorDay != 0 && currentDay > uint48(anchorDay + 1)) {
-            uint32 missedDays = uint32(currentDay - uint48(anchorDay) - 1);
+        if (anchorDay != 0 && currentDay > anchorDay + 1) {
+            uint32 missedDays = currentDay - anchorDay - 1;
             uint16 shields = questStreakShieldCount[player];
             if (shields != 0) {
                 uint32 used = missedDays > uint32(shields) ? uint32(shields) : missedDays;
@@ -1325,7 +1325,7 @@ contract DegenerusQuests is IDegenerusQuests {
     function _questSyncProgress(
         PlayerQuestState storage state,
         uint8 slot,
-        uint48 currentDay,
+        uint32 currentDay,
         uint24 questVersion
     ) private {
         uint24 currentDay24 = uint24(currentDay);
@@ -1348,7 +1348,7 @@ contract DegenerusQuests is IDegenerusQuests {
         PlayerQuestState memory state,
         DailyQuest memory quest,
         uint8 slot,
-        uint48 currentDay
+        uint32 currentDay
     ) private pure returns (bool) {
         if (quest.day == 0 || quest.day != currentDay) {
             return false;
@@ -1369,7 +1369,7 @@ contract DegenerusQuests is IDegenerusQuests {
         PlayerQuestState storage state,
         DailyQuest memory quest,
         uint8 slot,
-        uint48 currentDay
+        uint32 currentDay
     ) private view returns (bool) {
         if (quest.day == 0 || quest.day != currentDay) {
             return false;
@@ -1626,7 +1626,7 @@ contract DegenerusQuests is IDegenerusQuests {
         DailyQuest[QUEST_SLOT_COUNT] memory quests,
         uint8 slot,
         DailyQuest memory quest,
-        uint48 currentDay,
+        uint32 currentDay,
         uint256 mintPrice
     )
         private
@@ -1678,7 +1678,7 @@ contract DegenerusQuests is IDegenerusQuests {
         PlayerQuestState storage state,
         DailyQuest[QUEST_SLOT_COUNT] memory quests,
         uint8 slot,
-        uint48 currentDay,
+        uint32 currentDay,
         uint256 mintPrice
     )
         private
@@ -1760,8 +1760,8 @@ contract DegenerusQuests is IDegenerusQuests {
      * @param quests Memory array of active quests.
      * @return The current quest day (prefers slot 0 if both are set).
      */
-    function _currentQuestDay(DailyQuest[QUEST_SLOT_COUNT] memory quests) private pure returns (uint48) {
-        uint48 day0 = quests[0].day;
+    function _currentQuestDay(DailyQuest[QUEST_SLOT_COUNT] memory quests) private pure returns (uint32) {
+        uint32 day0 = quests[0].day;
         if (day0 != 0) return day0;
         return quests[1].day;
     }
